Add: support pulldown-cmark math.

+++ src/element_handler/element_util.rs
@@ -189,7 +189,7 @@ macro_rules! serialize_if_faithful {
         $num_attrs_allowed: expr
     ) => {
         if $handlers.options().translation_mode == $crate::options::TranslationMode::Faithful
-            && $element.attrs.len() > $num_attrs_allowed
+            && $element.attrs.len() as i64 > $num_attrs_allowed
         {
             return Some($crate::element_handler::HandlerResult {
                 content: $crate::element_handler::element_util::serialize_element(
